+++ packages/integration-sdk-core/src/types/entity.ts
@@ -3,17 +3,32 @@ import { PersistedObject } from './persistedObject';
 export type PrimitiveEntity = EntityCoreProperties &
   PrimitiveEntityAdditionalProperties;
 
+type EntityIdProperty = {
+  /**
+   * The natural identifier of the entity as provided by the data source API.
+   *
+   * Many APIs answer resources having a property representing the identity of
+   * the resource within the data source system. This value should be
+   * transferred to the entity's `id` property.
+   *
+   * In some cases an entity is known to a number of systems. The integrations
+   * that do not own the entity will use mapped relationships and provide the
+   * `id` value the data source maintains for the resource. The mapper will
+   * merge the values into a single Array.
+   *
+   * An identity value that is maintained as a number must be converted to a
+   * string. It is advisable to store the value in an additional property, such
+   * as `<resource>Id`, where `<resource>` reflects the type of resource
+   * represented by the entity (i.e. `domainId`, `userId`, etc.)
+   */
+  id?: string | string[];
+};
+
 type PrimitiveEntityAdditionalProperties = Record<
   string,
   PrimitiveEntityPropertyValue
-> & {
-  /**
-   * The natural identifier of the entity as provided by the data source API.
-   * Many APIs answer resources that each have an `id` property that should be
-   * transferred to this entity property.
-   */
-  id?: string;
-};
+> &
+  EntityIdProperty;
 
 type PrimitiveEntityPropertyValue =
   | Array<string | number | boolean>
@@ -34,14 +49,8 @@ interface EntityCoreProperties extends Omit<PersistedObject, '_class'> {
   _class: string | string[];
 }
 
-type EntityAdditionalProperties = Record<string, EntityPropertyValue> & {
-  /**
-   * The natural identifier of the entity as provided by the data source API.
-   * Many APIs answer resources that each have an `id` property that should be
-   * transferred to this entity property.
-   */
-  id?: string;
-};
+type EntityAdditionalProperties = Record<string, EntityPropertyValue> &
+  EntityIdProperty;
 
 type EntityPropertyValue = PrimitiveEntityPropertyValue | EntityRawData[];